Cache intermediates that require a GPU sync
Caches some intermediate keypoint values used while performing output prediction. This serves to avoid redundant GPU syncs in the inner loop.

Also replaces usage of unsqueeze_() with unsqueeze(), since the former changes dimension in-place instead of creating a view, which breaks caching.

I don't think this will have issues with multiple workers since all the cached data is never mutated (to my knowledge). But if it is mutated in ways I am not familiar with, it will not be multiple-worker safe